Source cleanups
