+++ gui/src/analysis/ray_volume_raster.h
@@ -9,6 +9,10 @@
 
 namespace analysis {
 
+/// Rasterize ray paths from a result set into a sparse 3D volume.
+///
+/// This is currently a work-in-progress analysis primitive used by the flux
+/// module for volume visualization.
 Result<analysis::SparseGrid3D<float>, QString>
 compute_ray_volume_raster(TaskControl&            promise,
                           unsigned                resolution,
