+++ gui/src/analysis/volume_to_mesh.h
@@ -8,6 +8,10 @@
 
 namespace analysis {
 
+/// Generate an isosurface mesh from a sparse scalar volume.
+///
+/// This is currently a work-in-progress marching-cubes style utility.
+/// TODO: Move to mc33
 Result<db::Mesh, QString>
 volume_to_mesh(TaskControl& output, SparseGrid3D<float> volume, float isoval);
-}
+} // namespace analysis
