+++ gui/src/database/entity.h
@@ -8,8 +8,9 @@
 
 namespace db {
 
-
-/// Wrapper type for Qt/QML interface
+/// Wrapper type for passing entt entities through Qt/QML.
+///
+/// Entity values are only meaningful inside the database that created them.
 struct Entity {
     Q_GADGET
     QML_VALUE_TYPE(db_entity);
@@ -17,6 +18,8 @@ struct Entity {
 
 public:
     Entity() = default;
+
+    /// Wrap an entt entity handle.
     Entity(entt::entity e) : value(e) { }
 
     entt::entity value = entt::null;
@@ -25,8 +28,10 @@ public:
 
     operator entt::entity() const { return value; }
 
+    /// Whether this wrapper contains a non-null entity handle.
     Q_INVOKABLE bool is_valid() const { return value != entt::null; }
 
+    /// Debug-friendly representation for QML and logs.
     Q_INVOKABLE QString debug_string() const {
         if (!is_valid()) return QStringLiteral("entity(null)");
         return QStringLiteral("entity(%1)").arg(entt::to_integral(value));
@@ -45,6 +50,7 @@ inline QDebug operator<<(QDebug debug, Entity const& c) {
 
 namespace std {
 template <>
+/// Allow db::Entity keys in unordered containers.
 struct hash<db::Entity> {
     std::size_t operator()(db::Entity e) const noexcept {
         return std::hash<entt::entity>()(e.value);
