+++ gui/src/database/mesh.h
@@ -7,12 +7,14 @@
 
 namespace db {
 
+/// Vertex format used by QQuick3DGeometry adapters.
 struct Vertex {
     glm::vec3 position;
     glm::vec3 normal;
     glm::vec2 uv; // quick geometry only supports floating
 };
 
+/// Triangle mesh exchanged between GUI analysis code and Quick3D adapters.
 struct Mesh {
     // using QVector for COW
     QVector<Vertex>     vertex;
