+++ gui/src/database/rootelementsmodel.h
@@ -32,6 +32,7 @@ public slots:
 
 // =============================================================================
 
+/// A model providing all element entities in a database.
 class AllElementsModel : public StructModelAdapter<EntityNamePair> {
     Q_OBJECT
 
@@ -49,21 +50,25 @@ public:
     explicit AllElementsModel(QObject* parent = nullptr);
     ~AllElementsModel() override = default;
 
+    /// Observe a database and rebuild the element list.
     void reset(Database* database);
 };
 
 
 // =============================================================================
 
-/// Only works with entity name pair + Element components
+/// Filter proxy for element models by material, geometry, and name.
+///
+/// Only works with source models that expose EntityNamePair rows backed by
+/// Element components.
 class InstanceSortFilter : public QSortFilterProxyModel {
     Q_OBJECT
 
     QPointer<Database> m_host;
 
-    Q_WRITABLE_PROPERTY(Entity, material_filter, {});
-    Q_WRITABLE_PROPERTY(Entity, geometry_filter, {});
-    Q_WRITABLE_PROPERTY(QString, name_filter, {});
+    Q_WRITABLE_PROPERTY(Entity, material_filter, { });
+    Q_WRITABLE_PROPERTY(Entity, geometry_filter, { });
+    Q_WRITABLE_PROPERTY(QString, name_filter, { });
 
     Q_READONLY_PROPERTY(bool, has_filter);
 
@@ -80,11 +85,17 @@ private slots:
 public:
     explicit InstanceSortFilter(QObject* parent = nullptr);
 
+    /// Observe a database so filters can resolve entity metadata.
     void reset(Database* database);
 
 public slots:
+    /// Clear the material-group filter.
     void clear_material();
+
+    /// Clear the geometry-group filter.
     void clear_geometry();
+
+    /// Clear all active filters.
     void clear_all_filters();
 
     // QSortFilterProxyModel interface
