+++ gui/src/database/simulationresult.h
@@ -21,15 +21,17 @@ namespace db {
 
 class Database;
 
+/// Inputs needed to translate library simulation results into GUI ray records.
 struct SimulationResultConversion {
     SolTrace::Result::SimulationResult&                                 result;
     SolTrace::Data::SimulationData const&                               data;
     std::unordered_map<SolTrace::Data::element_id, entt::entity> const& map;
 };
 
-// can probably make this a variant.
-// create and exit dont need entities
-
+/// Event type for one point along a traced ray path.
+///
+/// Some events, such as CREATE and EXIT, may not correspond to a database
+/// entity.
 enum class RayEventType : uint8_t {
     CREATE   = 1,
     ABSORB   = 2,
@@ -40,6 +42,7 @@ enum class RayEventType : uint8_t {
     UNKNOWN  = UINT8_MAX
 };
 
+/// One interaction/location along a ray path.
 struct RayEvent {
     glm::dvec3   location;
     glm::dvec3   direction;
@@ -47,11 +50,13 @@ struct RayEvent {
     RayEventType event;
 };
 
+/// Complete event chain for one simulated ray.
 struct RayRecord {
     uint64_t              id;
     std::vector<RayEvent> events;
 };
 
+/// GUI-owned simulation result plus lookup data used by analysis modules.
 class SimulationResult {
 public:
     SimulationResult();
@@ -69,12 +74,14 @@ public:
     std::unique_ptr<Database const> database;
 
     // TODO: Why is this fallible?
+    /// Convert a library result into GUI result records and entity mappings.
     static std::unique_ptr<SimulationResult>
     convert(SimulationResultConversion const&);
 };
 
 using SimulationResultPtr = std::shared_ptr<SimulationResult>;
 
+/// One saved result row exposed to QML.
 struct SimulationResultRecord {
     QString             name;
     QDateTime           when;
@@ -87,6 +94,7 @@ struct SimulationResultRecord {
                 SM_EXPOSE_RO(ray_count), );
 };
 
+/// QML-facing list model of completed simulation results.
 class SimulationResultModel
     : public StructModelAdapter<SimulationResultRecord> {
     Q_OBJECT
@@ -94,13 +102,23 @@ class SimulationResultModel
 public:
     explicit SimulationResultModel(QObject* parent = nullptr);
 
+    /// Return the result pointer at index, or null if index is invalid.
     SimulationResultPtr result_at(int index) const;
-    QString             name_at(int index) const;
+
+    /// Return the display name at index, or an empty string if invalid.
+    QString name_at(int index) const;
 
 public slots:
+    /// Append a completed result with generated display metadata.
     void append_result(SimulationResultPtr result);
+
+    /// Remove a result row.
     void remove_result(int index);
+
+    /// Rename a result row.
     void rename_result(int index, QString const& name);
+
+    /// Remove all result rows.
     void clear();
 };
