+++ gui/src/database/worldgeometrymodel.h
@@ -8,7 +8,10 @@
 
 namespace db {
 
-// TODO: push instance generation to a thread
+/// Quick3D instancing adapter for all elements that share one geometry group.
+///
+/// Instance data is rebuilt on the GUI thread today.
+/// TODO: push instance generation to a thread
 class InstancedElements : public QQuick3DInstancing {
     Q_OBJECT
 
@@ -39,19 +42,28 @@ private slots:
     entt::entity entity_at(int index);
 
 public slots:
-    // Select geometry models
+    /// Toggle selection for the instance at a rendered instance index.
     void toggle_selection(int index);
 
+    /// Override the display color for the instance at index.
     void set_color(int index, QColor color);
 
+    /// Set the default color for instances without a ColorComponent.
     void set_default_color(QColor color);
 
+    /// Material group assigned to the represented geometry group.
     db::Entity material_of_group();
+
+    /// Geometry group represented by this instancing adapter.
     db::Entity geometry_of_group();
 
+    /// Material group assigned to the instance at index.
     db::Entity material_of(int index);
+
+    /// Geometry group assigned to the instance at index.
     db::Entity geometry_of(int index);
 
+    /// Entity represented by the instance at index.
     db::Entity at(int index);
 
 public:
@@ -60,6 +72,7 @@ public:
                                QQuick3DObject* parent = nullptr);
 };
 
+/// One visible geometry group row consumed by the 3D scene.
 struct VisibleGroup {
     entt::entity                       geometry_group_entity;
     std::shared_ptr<InstancedElements> group_instances;
@@ -70,6 +83,7 @@ struct VisibleGroup {
                 SM_EXPOSE_RO(group_geometry), );
 };
 
+/// Model of visible geometry groups and their Quick3D adapters.
 class WorldGeometryModel : public StructModelAdapter<VisibleGroup> {
     Q_OBJECT
 
@@ -78,7 +92,7 @@ class WorldGeometryModel : public StructModelAdapter<VisibleGroup> {
     std::unordered_map<entt::entity, int> m_reverse;
     double                                m_surface_thickness = 0.05;
     unsigned                              m_subdivision_scale = 2;
-    QColor                                m_default_color = Qt::white;
+    QColor                                m_default_color     = Qt::white;
 
     QVector<VisibleGroup> rebuild_lists();
     void                  apply_surface_options(VisibleGroup const& group);
@@ -90,14 +104,20 @@ private slots:
     void group_removed(entt::entity);
 
 public slots:
+    /// Update default color on all visible instancing adapters.
     void set_default_color(QColor color);
+
+    /// Update generated surface thickness on all visible surface geometries.
     void set_surface_thickness(double thickness);
+
+    /// Update generated surface subdivision on all visible surface geometries.
     void set_subdivision_scale(unsigned scale);
 
 public:
     explicit WorldGeometryModel(QObject* parent = nullptr);
     virtual ~WorldGeometryModel() = default;
 
+    /// Observe a new database and rebuild the visible group list.
     void reset(Database* database);
 };
