+++ gui/src/job_control/job_run_thread.h
@@ -4,17 +4,20 @@
 
 #include "job_run_common.h"
 
+/// Simulation backend used by the threaded runner.
 enum class ThreadRunnerBackend {
     Native,
     Embree,
     Optix,
 };
 
+/// Configuration for one threaded simulation run.
 struct ThreadRunnerConfig {
     uint32_t            thread_count;
     ThreadRunnerBackend backend = ThreadRunnerBackend::Native;
 };
 
+/// Execute a simulation through the selected runner and report through promise.
 void execute_thread_runner(QPromise<SimResult>&      promise,
                            SimDataPtr                data,
                            ThreadRunnerConfig const& config);
