+++ gui/src/modules/database_module.h
@@ -13,18 +13,21 @@
 
 namespace SolTrace::GUI::App {
 
+/// One open database shown in the load/start workflow.
 struct DatabaseRecord {
     QPointer<db::Database> database;
 
     RECORD_META(DatabaseRecord, SM_EXPOSE_RO(database), );
 };
 
+/// Successful result from asynchronous database loading.
 struct LoadedFile {
     // TODO: Store this in the DB
     QString                       provenance = { };
     std::unique_ptr<db::Database> ptr;
 };
 
+/// Failed database load result packaged as a user notification.
 struct LoadFileFailed {
     ANotification notification;
 
@@ -34,6 +37,11 @@ struct LoadFileFailed {
 
 // using LoadResult = std::variant<LoadedFile, LoadFileFailed>;
 
+/// QML-facing controller for opening, saving, creating, and switching
+/// databases.
+///
+/// The model rows represent open databases. The module owns loaded database
+/// instances and exposes the selected one through current_database.
 class DatabaseModule : public StructModelAdapter<DatabaseRecord> {
     Q_OBJECT
     QML_ELEMENT
@@ -57,19 +65,34 @@ public:
     DatabaseModule(QObject* parent = nullptr);
 
 public slots:
+    /// Open a native file picker and load the selected SolTrace input file.
     bool open_file_dialog();
+
+    /// Open a native file picker and save the current database.
     bool save_current_dialog();
 
+    /// Load a database from a URL. name_override is used for display only.
     void load_url(QUrl, QString name_override = "");
+
+    /// Replace the current selection with a new blank database.
     void load_new();
 
+    /// Select the open database at model row index.
     bool set_current(int);
 
+    /// Save a specific open database to path.
     void save_db_at_index(int, QUrl);
+
+    /// Save current_database to path.
     void save_current(QUrl path);
 
+    /// Remove current_database from the open database list.
     void delete_current();
+
+    /// Add a new blank database to the open database list.
     void append_new(QString);
+
+    /// Clone a simulation result into an editable database.
     bool append_clone(db::SimulationResultPtr);
 
 signals:
