+++ gui/src/modules/flux_module.h
@@ -35,7 +35,7 @@ class FluxModule : public QObject {
     Q_READONLY_PROPERTY(bool, ray_volume_flux_in_progress);
     QOBJECT_READONLY_PROPERTY(db::QMLMesh, ray_iso_volume);
 
-    Q_WRITABLE_PROPERTY(db::Entity, current_entity, {});
+    Q_WRITABLE_PROPERTY(db::Entity, current_entity, { });
     Q_READONLY_PROPERTY(QString, current_entity_name);
     Q_READONLY_PROPERTY(analysis::BakedFluxMapStats, current_flux_stats);
 
@@ -43,13 +43,14 @@ class FluxModule : public QObject {
     Q_WRITABLE_PROPERTY(bool, show_other_geometry, false);
 
     // Hack
-    Q_WRITABLE_PROPERTY(QString, current_image, {});
+    Q_WRITABLE_PROPERTY(QString, current_image, { });
 
 private:
     void refresh_current_flux_stats();
 
 private slots:
-    void flux_map_ready(db::Entity, analysis::BakedFluxMapPtr, db::Database const*);
+    void
+    flux_map_ready(db::Entity, analysis::BakedFluxMapPtr, db::Database const*);
 
     void flux_vol_ready(QUuid const&, analysis::SparseGrid3D<float>);
     void flux_vol_failed(QUuid const&, QString);
@@ -61,12 +62,19 @@ public:
     explicit FluxModule(QQmlEngine*, QObject* parent = nullptr);
 
 public slots:
+    /// Set the result set used for all flux computations and scene models.
     void set_results(db::SimulationResultPtr);
+
+    /// Select the entity whose flux map/statistics are shown in the UI.
     void select_entity(db::Entity);
 
+    /// Generate a surface flux map for current_entity.
     void start_generate();
 
+    /// Generate a volumetric ray-density raster from the current result.
     void start_generate_volume_flux(unsigned resolution);
+
+    /// Generate an isosurface mesh from the current volumetric raster.
     void start_generate_isosurface(float value);
 
 signals:
